+++ stack-graphs/tests/it/can_find_root_partial_paths_in_database.rs
@@ -18,6 +18,7 @@ use stack_graphs::paths::ScopedSymbol;
 use stack_graphs::paths::SymbolStack;
 use stack_graphs::stitching::Database;
 use stack_graphs::stitching::SymbolStackKey;
+use stack_graphs::NoCancellation;
 
 use crate::test_graphs;
 
@@ -31,15 +32,17 @@ fn check_root_partial_paths(
     let mut paths = Paths::new();
     let mut partials = PartialPaths::new();
     let mut database = Database::new();
-    partials.find_all_partial_paths_in_file(graph, file, |graph, partials, path| {
-        if !path.is_complete_as_possible(graph) {
-            return;
-        }
-        if !path.is_productive(partials) {
-            return;
-        }
-        database.add_partial_path(graph, partials, path);
-    });
+    partials
+        .find_all_partial_paths_in_file(graph, file, &NoCancellation, |graph, partials, path| {
+            if !path.is_complete_as_possible(graph) {
+                return;
+            }
+            if !path.is_productive(partials) {
+                return;
+            }
+            database.add_partial_path(graph, partials, path);
+        })
+        .expect("should never be cancelled");
 
     let mut symbol_stack = SymbolStack::empty();
     for symbol in precondition.iter().rev() {