+++ tree-sitter-stack-graphs/src/test.rs
@@ -68,6 +68,8 @@ use std::path::Path;
 use std::path::PathBuf;
 use thiserror::Error;
 
+use crate::CancellationFlag;
+
 lazy_static! {
     static ref PATH_REGEX: Regex = Regex::new(r#"---\s*path:\s*([^\s]+)\s*---"#).unwrap();
     static ref ASSERTION_REGEX: Regex =
@@ -349,6 +351,7 @@ pub enum TestFailure {
         missing_lines: Vec<usize>,
         unexpected_lines: HashMap<String, Vec<Option<usize>>>,
     },
+    Cancelled(stack_graphs::CancellationError),
 }
 
 impl std::fmt::Display for TestFailure {
@@ -410,6 +413,7 @@ impl std::fmt::Display for TestFailure {
                 }
                 Ok(())
             }
+            Self::Cancelled(err) => write!(f, "{}", err),
         }
     }
 }
@@ -418,12 +422,15 @@ impl Test {
     /// Run the test. It is the responsibility of the caller to ensure that
     /// the stack graph has been constructed for the test fragments before running
     /// the test.
-    pub fn run(&mut self) -> TestResult {
+    pub fn run(
+        &mut self,
+        cancellation_flag: &dyn CancellationFlag,
+    ) -> Result<TestResult, stack_graphs::CancellationError> {
         let mut result = TestResult::new();
         for fragment in &self.fragments {
             for assertion in &fragment.assertions {
                 match assertion
-                    .run(&self.graph, &mut self.paths)
+                    .run(&self.graph, &mut self.paths, &cancellation_flag)
                     .map_or_else(|e| self.from_error(e), |v| Ok(v))
                 {
                     Ok(_) => result.add_success(),
@@ -431,7 +438,7 @@ impl Test {
                 }
             }
         }
-        result
+        Ok(result)
     }
 
     /// Construct a TestFailure from an AssertionError.
@@ -489,6 +496,7 @@ impl Test {
                     unexpected_lines,
                 })
             }
+            AssertionError::Cancelled(err) => Err(TestFailure::Cancelled(err)),
         }
     }