+++ tree-sitter-stack-graphs/tests/it/nodes.rs
@@ -9,15 +9,22 @@ use pretty_assertions::assert_eq;
 use stack_graphs::graph::StackGraph;
 use tree_sitter_graph::Variables;
 use tree_sitter_stack_graphs::LoadError;
+use tree_sitter_stack_graphs::NoCancellation;
 use tree_sitter_stack_graphs::StackGraphLanguage;
 
 fn build_stack_graph(python_source: &str, tsg_source: &str) -> Result<StackGraph, LoadError> {
-    let mut language =
+    let language =
         StackGraphLanguage::from_str(tree_sitter_python::language(), tsg_source).unwrap();
     let mut graph = StackGraph::new();
     let file = graph.get_or_create_file("test.py");
     let mut globals = Variables::new();
-    language.build_stack_graph_into(&mut graph, file, python_source, &mut globals)?;
+    language.build_stack_graph_into(
+        &mut graph,
+        file,
+        python_source,
+        &mut globals,
+        &NoCancellation,
+    )?;
     Ok(graph)
 }