+++ tree-sitter-stack-graphs/tests/it/test.rs
@@ -15,6 +15,7 @@ use std::path::PathBuf;
 use tree_sitter_graph::Variables;
 use tree_sitter_stack_graphs::test::Test;
 use tree_sitter_stack_graphs::LoadError;
+use tree_sitter_stack_graphs::NoCancellation;
 use tree_sitter_stack_graphs::StackGraphLanguage;
 
 lazy_static! {
@@ -57,10 +58,10 @@ fn build_stack_graph_into(
     python_source: &str,
     tsg_source: &str,
 ) -> Result<(), LoadError> {
-    let mut language =
+    let language =
         StackGraphLanguage::from_str(tree_sitter_python::language(), tsg_source).unwrap();
     let mut globals = Variables::new();
-    language.build_stack_graph_into(graph, file, python_source, &mut globals)?;
+    language.build_stack_graph_into(graph, file, python_source, &mut globals, &NoCancellation)?;
     Ok(())
 }
 
@@ -90,7 +91,9 @@ fn check_test(
         )
         .expect("Could not load stack graph");
     }
-    let results = test.run();
+    let results = test
+        .run(&NoCancellation)
+        .expect("should never be cancelled");
     assert_eq!(
         expected_successes,
         results.success_count(),