+++ tree-sitter-stack-graphs/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Library
+
+#### Changed
+
+- `StackGraphLanguage` instances can now safely be shared between threads. The `StackGraphLanguage::build_stack_graph_into` does not require a mutable instance anymore.
+- `StackGraphLanguage::build_stack_graph_into` and `loader::Loader::load_for_file` now supports cancellation by passing an instance of `CancellationFlag`. The `NoCancellation` type provides a noop implementation.
+- `test::Test::run` now supports cancellation by passing an instance of `CancellationFlag` and returns a `Result` indicating whether the test was canceled or not.
+
 ## 0.2.0 -- 2022-06-29
 
 Depend on `stack-graphs` version 0.9.